token-difr/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "token-difr"
version = "0.1.2"
description = "Verify LLM outputs using Gumbel-Max sampling verification"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["llm", "verification", "gumbel-max", "sampling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "datasets",
    "tqdm>=4.0.0",
    "python-dotenv",
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "modal>=0.64.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
vllm = [
    "vllm>=0.6.0",
]
all = [
    "vllm>=0.6.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "ipykernel==6.29.0"
]

[project.urls]
Homepage = "https://github.com/adamkarvonen/token-difr"
Repository = "https://github.com/adamkarvonen/token-difr"

[tool.hatch.build.targets.wheel]
packages = ["src/token_difr"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
